Add a Tile class
The Tile class stores information about a tile that can be placed in a
level. In addition to some metadata information useful to the user
(i.e., its ID, the name, and an optional description), the Tile class
also contains a reference to the Texture that should be used whenever
a visual representation of the Tile is needed.