Made some changes to the Sprite class
Don't allow the user to set the Position and Size after it has been
created and enforce the caller of the Sprite class to specify a
position and size upon construction. This is so that the SpriteSheet
class does not need to worry about the position and size of a Sprite
suddenly changing, and having to ensure that the updated position and
size is within range of the dimensions of the SpriteSheet.

In addition, add a SpriteSheet property that provides a reference to the
SpriteSheet that the sprite is associated with (which occurs when
SpriteSheet.Add(Sprite) is called). If a sprite has not been associated
with a spritesheet, then the property will be null.