Add a SpriteSheet class
This class is used to represent a collection a sprite that exists in a
single image. Given the path to the spritesheet, its width, and its
height, the class provides a way for the user to register the known
sprites that exists in the spritesheet into the spritesheet itself,
remove existing sprites, or checking the existence of a sprite.

This class implements the ICollection<Sprite> interface.